updated middleware structure
